Maryia/refactoring
